+++ localflavor/au/models.py
@@ -1,4 +1,4 @@
-from django.db.models.fields import CharField
+from django.db.models import CharField
 from django.utils.translation import ugettext_lazy as _
 
 from . import forms
@@ -18,6 +18,12 @@ class AUStateField(CharField):
         kwargs['max_length'] = 3
         super(AUStateField, self).__init__(*args, **kwargs)
 
+    def deconstruct(self):
+        name, path, args, kwargs = super(AUStateField, self).deconstruct()
+        del kwargs['choices']
+        del kwargs['max_length']
+        return name, path, args, kwargs
+
 
 class AUPostCodeField(CharField):
     """
@@ -30,6 +36,11 @@ class AUPostCodeField(CharField):
     def __init__(self, *args, **kwargs):
         kwargs['max_length'] = 4
         super(AUPostCodeField, self).__init__(*args, **kwargs)
+
+    def deconstruct(self):
+        name, path, args, kwargs = super(AUPostCodeField, self).deconstruct()
+        del kwargs['max_length']
+        return name, path, args, kwargs
 
     def formfield(self, **kwargs):
         defaults = {'form_class': forms.AUPostCodeField}
@@ -47,6 +58,11 @@ class AUPhoneNumberField(CharField):
     def __init__(self, *args, **kwargs):
         kwargs['max_length'] = 20
         super(AUPhoneNumberField, self).__init__(*args, **kwargs)
+
+    def deconstruct(self):
+        name, path, args, kwargs = super(AUPhoneNumberField, self).deconstruct()
+        del kwargs['max_length']
+        return name, path, args, kwargs
 
     def formfield(self, **kwargs):
         defaults = {'form_class': forms.AUPhoneNumberField}