+++ localflavor/mk/models.py
@@ -1,4 +1,4 @@
-from django.db.models.fields import CharField
+from django.db.models import CharField
 from django.utils.translation import ugettext_lazy as _
 
 from .mk_choices import MK_MUNICIPALITIES
@@ -17,6 +17,11 @@ class MKIdentityCardNumberField(CharField):
         kwargs['max_length'] = 8
         super(MKIdentityCardNumberField, self).__init__(*args, **kwargs)
 
+    def deconstruct(self):
+        name, path, args, kwargs = super(MKIdentityCardNumberField, self).deconstruct()
+        del kwargs['max_length']
+        return name, path, args, kwargs
+
     def formfield(self, **kwargs):
         defaults = {'form_class': MKIdentityCardNumberFormField}
         defaults.update(kwargs)
@@ -34,6 +39,12 @@ class MKMunicipalityField(CharField):
         kwargs['choices'] = MK_MUNICIPALITIES
         kwargs['max_length'] = 2
         super(MKMunicipalityField, self).__init__(*args, **kwargs)
+
+    def deconstruct(self):
+        name, path, args, kwargs = super(MKMunicipalityField, self).deconstruct()
+        del kwargs['choices']
+        del kwargs['max_length']
+        return name, path, args, kwargs
 
 
 class UMCNField(CharField):
@@ -55,6 +66,11 @@ class UMCNField(CharField):
     def __init__(self, *args, **kwargs):
         kwargs['max_length'] = 13
         super(UMCNField, self).__init__(*args, **kwargs)
+
+    def deconstruct(self):
+        name, path, args, kwargs = super(UMCNField, self).deconstruct()
+        del kwargs['max_length']
+        return name, path, args, kwargs
 
     def formfield(self, **kwargs):
         defaults = {'form_class': UMCNFormField}