+++ tests/test_generic/tests.py
@@ -450,3 +450,11 @@ class EANTests(TestCase):
 
         for value in invalid:
             self.assertRaisesMessage(ValidationError,  error_message, validator, value)
+
+    def test_deconstruct_methods(self):
+        # test_instance must be created with the non-default options.
+        test_instance = IBANField(include_countries=('NL', 'BE'), use_nordea_extensions=True)
+        name, path, args, kwargs = test_instance.deconstruct()
+        new_instance = IBANField(*args, **kwargs)
+        for attr in ('include_countries', 'use_nordea_extensions'):
+            self.assertEqual(getattr(test_instance, attr), getattr(new_instance, attr))